fix:service list command
