--- cmd/service/cmd.go
@@ -1,38 +1,0 @@
-// Copyright 2022 API7.ai, Inc
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package services
-
-import (
-	"github.com/spf13/cobra"
-
-	"github.com/api7/cloud-cli/internal/output"
-	"github.com/api7/cloud-cli/internal/persistence"
-)
-
-// NewCommand creates the service sub-command object.
-func NewCommand() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "service [COMMAND] [ARGS...]",
-		Short: "service management",
-		PersistentPreRun: func(cmd *cobra.Command, args []string) {
-			if err := persistence.CheckConfigurationAndInitCloudClient(); err != nil {
-				output.Errorf(err.Error())
-			}
-		},
-	}
-
-	cmd.AddCommand(newListCommand())
-	return cmd
-}
