+++ main.go
@@ -24,7 +24,6 @@ import (
 	"github.com/api7/cloud-cli/cmd/debug"
 	"github.com/api7/cloud-cli/cmd/deploy"
 	"github.com/api7/cloud-cli/cmd/resource"
-	"github.com/api7/cloud-cli/cmd/service"
 	"github.com/api7/cloud-cli/cmd/stop"
 	"github.com/api7/cloud-cli/internal/options"
 	"github.com/api7/cloud-cli/internal/utils"
@@ -46,7 +45,6 @@ func newCommand() *cobra.Command {
 	cmd.AddCommand(debug.NewCommand())
 	cmd.AddCommand(config.NewCommand())
 	cmd.AddCommand(resource.NewCommand())
-	cmd.AddCommand(services.NewCommand())
 
 	return cmd
 }
